+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 ==================
 New Functionalities:
 * [OE-96](https://openlmis.atlassian.net/browse/OE-96): Added quantity unit toggle component
+New Functionalities:
+* [OE-108](https://openlmis.atlassian.net/browse/OE-108): Added quantity unit input component
 
 7.2.13 / 2025-03-31
 ==================